// Frontend/src/app/api/models/candidate-application-history-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CandidateApplicationHistoryDto {
  canWithdraw?: boolean;
  cvUrl?: string;
  id: number;
  jobOfferId?: number;
  jobTitle?: string;
  jobType?: 'PER' | 'PATS' | 'CONTRACTUEL';
  lastUpdateDate?: string;
  status?: string;
  submitDate?: string;
  typeContrat?: 'CDD' | 'CDI' | 'STAGE';
}